Only allow users to delete their own search queries. Also, fix an issue where 'testDeleteSearchQuery' was not actually testing what it purported to be testing

+++ src/java/com/app/controller/SearchQueryController.java
@@ -133,13 +133,18 @@ public class SearchQueryController {
 			for (String searchQueryId : searchQueryIds) {
 				int searchQueryIdInteger = Integer.parseInt(searchQueryId);
 
-				SearchQueryUtil.deleteSearchQuery(searchQueryIdInteger);
-
-				SearchResultUtil.deleteSearchQueryResults(
+				SearchQuery searchQuery = SearchQueryUtil.getSearchQuery(
 					searchQueryIdInteger);
 
-				SearchQueryPreviousResultUtil.deleteSearchQueryPreviousResults(
-					searchQueryIdInteger);
+				if (searchQuery.getUserId() == UserUtil.getCurrentUserId()) {
+					SearchQueryUtil.deleteSearchQuery(searchQueryIdInteger);
+
+					SearchResultUtil.deleteSearchQueryResults(
+						searchQueryIdInteger);
+
+					SearchQueryPreviousResultUtil.deleteSearchQueryPreviousResults(
+						searchQueryIdInteger);
+				}
 			}
 		}
